requirements.txt
psycopg[binary]
pydantic
python-dotenv
mcp
fastapi
uvicorn



